+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Note: Breaking changes between versions are indicated by "💥".
 
 ## Unreleased
 
-- 💥[Improvement] Get rid of the `indexcourses` command (#269)
+- 💥[Improvement] Get rid of the `indexcourses` and `portainer` command (#269)
 
 ## 3.8.0 (2019-11-22)